Tratado erro de violação de chave estrangeira após exclusão do diário de avaliações Refs #388

+++ app/controllers/avaliations_controller.rb
@@ -82,10 +82,14 @@ class AvaliationsController < ApplicationController
 
   def destroy
     authorize resource
+    if resource.recovery_diary_record
+      flash[:alert] = t("flash.avaliations.destroy.foreign_key_violation")
+      redirect_to avaliations_path
+    else
+      resource.destroy
 
-    resource.destroy
-
-    respond_with resource, location: avaliations_path
+      respond_with resource, location: avaliations_path
+    end
   end
 
   def history
